// src/app/mill_management_system/model/withdrawal/mill-withdrawal-details.model.ts
export class MillWithdrawalDetails {
    serialNo: number | undefined;
    withdrawalDate: any | undefined;
    noOfBags: number = 0;
    quantity: number = 0;
    transporterName: string | undefined;
    vehicleNo: string | undefined;
    driverName: string | undefined;
    driverMobNo: string | undefined;
    deliveryChallanNo: string | undefined;
}